Reorganized contexts into its own submodule.
